refactor: Use usize in route_back_cache.rs instead of doing casting to u64 (#5314)
In `route_back_cache.rs` we have been casting `usize` to `u64` a few times. We even had `Size` type defined as `u64`.
Let's clean up this code by changing types representing size of hashmaps to native `usize`.

+++ chain/network/src/routing/routing.rs
@@ -26,9 +26,9 @@ use crate::routing::utils::cache_to_hashmap;
 use crate::PeerInfo;
 
 const ANNOUNCE_ACCOUNT_CACHE_SIZE: usize = 10_000;
-const ROUTE_BACK_CACHE_SIZE: u64 = 100_000;
+const ROUTE_BACK_CACHE_SIZE: usize = 100_000;
 const ROUTE_BACK_CACHE_EVICT_TIMEOUT: Duration = Duration::from_millis(120_000);
-const ROUTE_BACK_CACHE_REMOVE_BATCH: u64 = 100;
+const ROUTE_BACK_CACHE_REMOVE_BATCH: usize = 100;
 const PING_PONG_CACHE_SIZE: usize = 1_000;
 const ROUND_ROBIN_MAX_NONCE_DIFFERENCE_ALLOWED: usize = 10;
 const ROUND_ROBIN_NONCE_CACHE_SIZE: usize = 10_000;
